+++ CHANGELOG.md
@@ -35,7 +35,7 @@ From [GitHub PR #1699](https://github.com/microsoft/vsts-extension-retrospective
 ## v1.92.51
 
 * Fix: improve readability of retrospective summary percentages by adjusting background styling. From [GitHub PR #1552](https://github.com/microsoft/vsts-extension-retrospectives/pull/1552)
-* Feat: update AdjustIcon SVG for improved design and functionality. From [GitHub PR #1554](https://github.com/microsoft/vsts-extension-retrospectives/pull/1554)
+* Feat: update SVG for adjust icon (renamed focus icon) for improved design and functionality. From [GitHub PR #1554](https://github.com/microsoft/vsts-extension-retrospectives/pull/1554)
 * Fix: update Retrospective Wiki URL to the correct domain. From [GitHub PR #1556](https://github.com/microsoft/vsts-extension-retrospectives/pull/1556)
 
 ## v1.92.50